factored out scraping logic
